+++ src/projects/modules/containers/bmff/fmp4_packager/fmp4_packager.h
@@ -21,7 +21,7 @@ namespace bmff
 	public:
 		struct Config
 		{
-			uint32_t chunk_duration_ms = 1000;
+			double chunk_duration_ms = 1000.0;
 		};
 
 		FMP4Packager(const std::shared_ptr<FMP4Storage> &storage, const std::shared_ptr<const MediaTrack> &media_track, const std::shared_ptr<const MediaTrack> &data_track, const Config &config);
@@ -42,7 +42,7 @@ namespace bmff
 		std::shared_ptr<bmff::Packager::Samples> GetDataSamples(int64_t start_timestamp, int64_t end_timestamp);
 
 		bool StoreInitializationSection(const std::shared_ptr<ov::Data> &segment);
-		bool AppendMediaChunk(const std::shared_ptr<ov::Data> &chunk, uint64_t start_timestamp, uint32_t duration_ms, bool independent);
+		bool AppendMediaChunk(const std::shared_ptr<ov::Data> &chunk, int64_t start_timestamp, double duration_ms, bool independent);
 
 		std::shared_ptr<const MediaPacket> ConvertBitstreamFormat(const std::shared_ptr<const MediaPacket> &media_packet);
